refactor events and serialization
